+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -5,7 +5,17 @@
 ### Breaking Changes
 
 - Removed `Azure::Storage::Blobs::PackageVersion`.
-- Renamed `GetUserDelegationKeyOptions::startsOn` to `GetUserDelegationKeyOptions::StartsOn`.
+- Renamed `GetUserDelegationKeyOptions::startsOn` to `StartsOn`.
+- Replaced all paginated collection functions that have the SinglePage suffix with pageable functions returning a `PagedResponse<T>`-derived type. The options are also renamed accordingly.
+  - `BlobServiceClient::ListBlobContainers()`.
+  - `BlobServiceClient::FindBlobsByTags()`.
+  - `BlobContainerClinet::ListBlobs()`.
+  - `BlobContainerClient::ListBlobsByHierarchy()`.
+  - `PageBlobClient::GetPageRanges()`.
+  - `PageBlobClient::GetPageRangesDiff()`.
+  - `PageBlobClient::GetManagedDiskPageRangesDiff()`.
+- Renamed `FilterBlobItem` to `TaggedBlobItem`.
+  - `FindBlobsByTags()` now returns `FindBlobsByTagsPagedResponse` and the field `FindBlobsByTagsSinglePageResult::Items` was renamed to `FindBlobsByTagsPagedResponse::TaggedBlobs`.
 
 ## 12.0.0-beta.9 (2021-03-23)