+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -5,7 +5,12 @@
 ### Breaking Changes
 
 - Removed `Azure::Storage::Files::DataLake::PackageVersion`.
-- Renamed `GetUserDelegationKeyOptions::startsOn` to `GetUserDelegationKeyOptions::StartsOn`.
+- Renamed `GetUserDelegationKeyOptions::startsOn` to `StartsOn`.
+- Replaced all paginated collection functions that have the SinglePage suffix with pageable functions returning a `PagedResponse<T>`-derived type. The options are also renamed accordingly.
+  - `DataLakeServiceClient::ListFileSystems()`.
+  - `DataLakeFileSystemClient::ListPaths()`.
+  - `DataLakeDirectoryClient::ListPaths()`.
+- Removed `DataLakePathClient::SetAccessControlListRecursiveSinglePage()`, `UpdateAccessControlListRecursiveSinglePage()` and `RemoveAccessControlListRecursiveSinglePage()`.
 
 ### Bug Fixes